+++ build/wix/clean.bat
@@ -1,8 +1,8 @@
 @ECHO OFF
 
-del *.wixobj
-del *.wixpdb
-del *.log
-del *.msi
-del /Q subdirs\*.*
+if exist *.wixobj del *.wixobj
+if exist *.wixpdb del *.wixpdb
+if exist *.log del *.log
+if exist *.msi del *.msi
+if exist subdirs\*.* del /Q subdirs\*.*
 type NUL > subdirs\_EMPTY_
